refactor: getUserMatchRecords 함수 파일 분리
